Prep for other implementations
+ split general functions and defs into storage.go
+ create skeleton for getter.go
+ remove old empty test for appengine native generator
