+++ pkg/storage/gcp/getter.go
@@ -1,0 +1,7 @@
+package gcp
+
+import "github.com/gomods/athens/pkg/storage"
+
+func (s *Storage) Get(module, version string) (*storage.Version, error) {
+	return nil, nil
+}
